requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# LangGraph
langgraph>=0.2.0
langchain-core>=0.3.0

# IBM Services
ibm-watson>=8.0.0
ibm-watsonx-ai>=1.0.0
ibm-cloud-sdk-core>=3.18.0

# Database
asyncpg>=0.29.0
pgvector>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
